Extract `BuildGraphData` interface
to make it easier to understand what it is responsible for.

Bug: n/a
Test: n/a

AOSP: ba4956357370ed2f5d626a7ced5ee89fb695131e
