+++ querysync/javatests/com/google/idea/blaze/qsync/project/BuildGraphDataTest.java
@@ -47,7 +47,7 @@ public class BuildGraphDataTest {
 
   @Test
   public void pathToLabel() {
-    BuildGraphData.Builder builder = BuildGraphData.builder();
+    BuildGraphDataImpl.Builder builder = BuildGraphDataImpl.builder();
     builder.sourceFileLabelsBuilder()
       .add(Label.of("//:BUILD"))
       .add(Label.of("//nested:BUILD"))
@@ -78,7 +78,7 @@ public class BuildGraphDataTest {
 
   @Test
   public void sourceFileToLabel() {
-    BuildGraphData.Builder builder = BuildGraphData.builder();
+    BuildGraphDataImpl.Builder builder = BuildGraphDataImpl.builder();
     builder.sourceFileLabelsBuilder()
       .add(Label.of("//:BUILD"))
       .add(Label.of("//nested:BUILD"))
